refactor: 리뷰 의견 받아서 추상화 낮게 적용
